pan.bd {pan}	R Documentation

Imputation of multivariate panel or cluster data

Description

Implementation of pan() that restricts the covariance matrix for the random effects to be block-diagonal. This function is identical to pan() in every way except that psi is now characterized by a set of r matrices of dimension q x q.

Usage

pan.bd(y, subj, pred, xcol, zcol, prior, seed, iter=1, start)

Arguments

y
See description for pan().

subj
See description for pan().

pred
See description for pan().

xcol
See description for pan().

zcol
See description for pan().

prior
Same as for pan() except that the hyperparameters for psi have new dimensions. The hyperparameter c is now a vector of length r, where c[j] contains the prior degrees of freedom for the jth block portion of psi (j=1,...,r). The hyperparameter Dinv is now an array of dimension c(q,q,r), where Dinv[,,j] contains the prior scale matrix for the jth block portion of psi (j=1,...,r).

seed
See description for pan().

iter
See description for pan().

start
See description for pan().

Value

A list with the same components as that from pan(), with two minor differences: the dimension of "psi" is now (q x q x r x "iter"), and the dimension of "last$psi" is now (q x q x r).

[Package pan version 1.9 Index]